Move event calls from frmMain into events.cpp
